backend/requirements.txt
Flask>=3.0.0
Flask-CORS>=4.0.0
Werkzeug>=3.0.1
watchdog>=3.0.0
Pillow>=10.0.0
PyPDF2>=3.0.0
openai>=1.0.0
anthropic>=0.18.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
